refactor error module
